+++ src/components/userList/index.tsx
@@ -1,5 +1,6 @@
 import React, {Component} from 'react';
 import {View, Text} from 'react-native';
+import styles from './styles';
 
 export default class UserList extends Component {
   render() {
