+++ src/components/userList/styles.tsx
@@ -1,0 +1,3 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({});
